Add files via upload
script to store stat() and lfs getstripe information for all files in a specified directory, as well as a df overview. 
Worked fine for ~99.99% of files, but sometimes a path would not be correctly created or getstripe information be empty for unknown reasons. Those cases are simply skipped right now.
Also quite slow for large directories